% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addAnchor.R
\name{addAnchor}
\alias{addAnchor}
\title{Dodaje kolumny grupujące wartości dla zadań kotwiczących.}
\usage{
addAnchor(df, anchor_mat, test_ver = "wersja")
}
\arguments{
\item{df}{data.frame, do którego dodane mają być kolumny.}

\item{anchor_mat}{macierz wskazująca odpowiadające sobie zadania w
poszczególnych wersjach.}

\item{test_ver}{character z nazwą kolumny w \code{df}, w której znajduje się
 informacja o rozwiązywanej wersji testu. Alternatywnie wektor z taką
 informacją (o długości równej liczbie wierszy w \code{df}). Domyślnie
 "wersja".}
}
\description{
Dodaje do zbioru danych kolumny, w których gromadzi wartości dla zadań
kotwiczących zdefiniowanych w \code{anchor_mat}.
}

