% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/majkTools.R
\docType{data}
\name{context_dat}
\alias{context_dat}
\title{Dane kontekstowe dla 35 uczniów w formacie .dat (ConQuest2).}
\format{Character zawierający 35 wierszy z kolumnami:
\describe{
  \item{id_szk}{znaki od 1 do 4}
  \item{id_kl}{znaki od 5 do 5}
  \item{id_ucz}{znaki od 6 do 9}
  \item{plec}{znaki od 10 do 10}
}}
\usage{
context_dat
}
\description{
Dane kontekstowe dla 35 uczniów, zawierające różne fikcyjne identyfikatory,
oryginalnie zapisane w formacie pliku danych ConQuesta 2 (plik tekstowy,
bez separatorów pól, jeden wiersz = jeden uczeń, kolumny o róznej szerokości)
i wczytane za pomocą funkcji \code{\link[base]{readLines}}.
}
\keyword{datasets}

