% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertScores.R
\encoding{UTF-8}
\name{convertScores}
\alias{convertScores}
\title{Przekształca wyniki z ConQuesta 2}
\usage{
convertScores(scores = "show.cases", write_to = NULL, pv = 5,
  dat_file = NULL, ...)
}
\arguments{
\item{scores}{Character ze ścieżką do outputu z ConQuesta2 dla
polecenia show.cases. Alternatywnie character z wynikiem działania funkcji
\code{\link[base]{readLines}} dla tego pliku.}

\item{write_to}{Character z nazwą (ścieżką) pliku wynikowego w formacie csv.
Jeżeli NULL, zwraca data.frame z wynikiem.}

\item{pv}{Integer okreslający liczbę wygenerowanych pv. Potrzebny do
określenia długości tablicy z wynikami dla jednego ucznia w \code{scores}.}

\item{dat_file}{Character ze ścieżką do pliku z danymi, z których ConQuest
wygenerował output (opcjonalny). Alternatywnie character z wynikiem działania funkcji
\code{\link[base]{readLines}} dla tego pliku.}

\item{...}{Definicja pozycji kolumn w pliku dat_file, które mają być
 dodane do wyników (patrz sekcja Details).}
}
\value{
Zbiór danych w postaci tabelarycznej z oszacowaniami (EAP, SD i PV) w
kolumnach. Uporządkowanie uczniów jest identyczne z tym w pliku
\code{scores}. Jeżeli podana została ścieżka zapisu (\code{write_to}), zbiór
jest zapisany do pliku. Domyślnie zwraca data.frame.

Jeżeli podano ścieżkę do pliku .dat (\code{dat_file}) i zdefiniowano pozycje
kolumn dodatkowych z tego pliku, zbiór danych jest powiększony o te
kolumny (patrz sekcja Details).

W przypadku, gdy wyniki są wyliczone z modelu wielowymiarowego, kolumny z
oszacowaniami dla kolejnych \code{N} wymiarów mają przedrostek \code{DimN_}.
}
\description{
Przekształca wyniki z ConQuesta 2 otrzymane w wyniku działania polecenia
\code{show cases} do postaci data.frame i umożliwia ich zapis na dysku.
Dodatkowo umożliwia dołączenie kolumn (np. identyfikatorów) z pliku .dat,
z którego wyliczone zostały wyniki w ConQueście.
}
\details{
Funkcja wymaga pakietu
\href{http://cran.r-project.org/web/packages/plyr/index.html}{plyr}
do działania.

Definicja pozycji kolumn powinna przyjmować następującą strukturę:
 \code{varName = c(n, k)}. \code{varName} określa nazwę pod jaką kolumna
 zostanie dodana do zbioru. \code{n, k} to liczby określające, odpowiednio,
 numer startowego i końcowego indeksu w wierszu w pliku tekstowym
 \code{dat_file}. Jeżeli kolumna ma szerokość równą jeden, należy dwukrotnie
 podać pozycję tej kolumny (np. \code{id_ucz = c(2, 2)}). Kolejne definicje
 kolumn należy dodawać po przecinku.

Zdefiniowane kolumny zostaną dodane na początku wynikowego zbioru danych w
kolejności w jakiej zostały zdefiniowane.
}

