% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getKey.R
\name{getKey}
\alias{getKey}
\title{Generuje z codebooka klucz wartości potrzebnych do rekodowania.}
\usage{
getKey(cdbook, vars, default_key = c("1", "2"), mcq_key = NULL,
  mcq_val = 4, na_codes = NULL)
}
\arguments{
\item{cdbook}{data.frame z codebookiem z weryfikatora.}

\item{vars}{string z wyrażeniem regularnym lub wektor tekstowy z nazwami
kolumn z zadaniami.}

\item{default_key}{wektor tekstowy określający wartości w bazie do zrekodowania,
które mają nie być rekodowane na 0 (zostana pozostawione). Dotyczy tylko
zadań nie określonych jako zamknięte w \code{mcq_val} (patrz Details).
Domyślnie \code{c("1", "2")}.}

\item{mcq_key}{wektor określający wartości dla zadań zamkniętych, którym po
zrekodowaniu przypisać wartość 1, a nie 0. Liczba kodów musi
zgadzać się z liczbą wykrytych zadań zamkniętych (patrz Details).}

\item{mcq_val}{integer określający wartość, która występuje tylko dla zadań
zamkniętych (jednoznacznie je wskazuje). Domyślnie 4.}

\item{na_codes}{wektor określający wartości w bazie, które mają być
 zrekodowane jako NA, a nie 0. Dotyczy wszystkich zmiennych określonych w
 \code{vars}.}
}
\value{
Lista o następujących elementach:
items     wektor tekstowy z nazwami zmiennych pasujących do 'vars'.
values    lista wartości zdefiniowanych dla 'items'.
recodes   lista wartości, na które mają być zrekodowane 'values'.
open      wektor tekstowy z nazwami zmiennych określonych jako otwarte
          (tj. nie-zamknięte).
mcq       wektor tekstowy z nazwami zmiennych określonych jako zamknięte.
open_key  wykorzystany 'open_key'.
mcq_key   wykorzystany 'mcq_key' (o ile został podany).
}
\description{
Generuje z codebooka klucz wartości potrzebnych do rekodowania.
}
\details{
Ważne, żeby nazwy zmiennych w \code{cdbook} były zdefiniowane w kolumnie
'Nazwa( zmiennej)', a etykiety wartości w kolumnie 'Etykiety( wartości)'.
Część nie w nawiasie stanowi zapytanie odnajdujące odpowiednie kolumny.
Wyrażenia 'Nazwa' oraz 'Etykiety' muszą więc identyfikować dokładnie po
jednej zmiennej (patrz reformat()).

Zakłada się, że wartości są oddzielone od swoich etykiet za pomocą
dwukropka ':', bez spacji po obu stronach.

Jeżeli \code{mcq_key} nie jest \code{NULL}, to w trakcie generowania klucza
zakłada się, że zmienne, które posiadają etykietę dla wartości określonej w
\code{mcq_val}, są zadaniami zamkniętymi.
}

