% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotIRT.R
\name{getProbIRT}
\alias{getProbIRT}
\title{Zwraca prawdopodobieństwo poprawnej odpowiedzi na zadanie.}
\usage{
getProbIRT(theta, b = 0, a = 1, c = 0, d = 1, const = 1.7,
  logit = FALSE)
}
\arguments{
\item{theta}{numeric, punkty na skali, dla których
ma być wyliczona wartość funkcji.}

\item{b}{numeric, trudnoścć zadania, domyślnie 0.}

\item{a}{numeric, dyskryminacja zadania, domyślnie 1.}

\item{c}{numeric, dolna asymptota (parametr pseudo-zgadywania),
w zakresie \code{c(0, 1)}, domyślnie 0.}

\item{d}{numeric, górna asymptota (parametr pseudo-nieuważności),
w zakresie \code{c(0, 1)}, domyślnie 1.}

\item{const}{numeric z określeniem stałej zbliżającej skalę do rozkładu
normalnego.}

\item{logit}{logical, czy wartość funkcji ma być przedstawiona na skali logitowej?
 Domyślnie \code{FALSE} (tj. skala prawdopodobieństwa).}
}
\value{
Numeric, o tej samej długości co \code{theta}, którego elementy
 odpowiadają prawdopodobieństwu poprawnej odpowiedzi na zadanie przy
 określonych parametrach. Jeżeli \code{logit = TRUE}, wartość jest wyrażona
 w logitach (tj. odpowiada logarytmowi naturalnemu ilorazu szans).
}
\description{
Zwraca prawdopodobieństwo poprawnej odpowiedzi na zadanie przy
 określonych parametrach zadania.
}

