% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hasAnything.R
\name{hasAnything}
\alias{hasAnything}
\title{Znajdź wiersze, które coś mają}
\usage{
hasAnything(df)
}
\arguments{
\item{data.frame}{lub macierz z wierszami do sprawdzenia}
}
\value{
Wektor logiczny mówiący, które wiersze posiadają wartość w
którejkolwiek kolumnie z \code{df}.
}
\description{
Zwraca wektor logiczny z informacją, które wiersze z data.frame lub macierzy, które
przynajmniej na jednej kolumnie posiadają wartość (tj. nie NA lub NaN).
}

