\name{majkTools-package}
\alias{majkTools-package}
\alias{majkTools}
\docType{package}
\title{
Różniaste funkcje, które okazały się całkiem przydatne
}
\description{
Pakiet zbiera funkcje, które w miedzyczasie napisałem, a które
    miałem okazję wykorzystać więcej niż jeden raz:). Póki co, znajduje się tu 
    tylko funkcja wspomagające obróbkę danych z ConQuesta2, ale pojawią też
    inne...
}
\details{
\tabular{ll}{
Package: \tab majkTools\cr
Type: \tab Package\cr
Version: \tab 0.0.0.9000\cr
Date: \tab 2015-02-11\cr
License: \tab GPL (>= 2)\cr
}

}