% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotIRT.R
\name{plotICC}
\alias{plotICC}
\title{Rysuje ICC dla podanych parametrów.}
\usage{
plotICC(b = 0, a = 1, c = 0, d = 1, const = 1.7, theta = c(-3, 3),
  logit = FALSE, ...)
}
\arguments{
\item{b}{numeric, trudnoścć zadania, domyślnie 0.}

\item{a}{numeric, dyskryminacja zadania, domyślnie 1.}

\item{c}{numeric, dolna asymptota (parametr pseudo-zgadywania),
w zakresie \code{c(0, 1)}, domyślnie 0.}

\item{d}{numeric, górna asymptota (parametr pseudo-nieuważności),
w zakresie \code{c(0, 1)}, domyślnie 1.}

\item{const}{numeric z określeniem stałej zbliżającej skalę do rozkładu
normalnego.}

\item{theta}{dwuelementowy numeric z określeniem przedziału skali, dla którego
ma być wyliczona wartość funkcji. Domyślnie \code{c(-3, 3)}.}

\item{logit}{logical, czy wartość funkcji ma być przedstawiona na skali logitowej?
Domyślnie \code{FALSE} (tj. skala prawdopodobieństwa).}

\item{...}{parametry przekazane do funkcji \code{\link[graphics]{plot}}.}
}
\description{
Rysuje wykresy z krzywą charaktersytyczną dla zadania (ICC), dla
 podanych parametrów zadania.
}

