% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readKey.R
\name{readKey}
\alias{readKey}
\title{Wczytuje plik ze zdefiniowanym kluczem}
\usage{
readKey(file, ...)
}
\arguments{
\item{file}{character wskazujący plik .csv z kluczem.}

\item{...}{inne argumenty przekazane do \code{\link[utils]{read.csv2}}.}
}
\value{
Lista z nazwami odpowiadającymi zadaniom, której elementami są data.frame'y
z kolumnami "values" i "recode" z klucza.
}
\description{
Wczytuje klucz zapisany w pliku tekstowym za pomocą funkcji
\code{\link[utils]{read.csv2}}.

W pliku, nazwy zadań powinny byc zdefiniowane w kolumnie 'item' i być
powtórzone tyle razy, ile wartości zdefiniowano dla danego zadania do
zrekodowania. W kolumnie "value" powinny znajdować się wartości zdefiniowane
dla zadania w oryginalnej bazie danych. W kolumnie "recode" powinny
znajdować się odpowiadające im wartości docelowe.
}

