% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recodeData.R
\name{recodeData}
\alias{recodeData}
\title{Rekoduje zbiór danych zgodnie z kluczem.}
\usage{
recodeData(raw_df, key_list)
}
\arguments{
\item{raw_df}{data.frame z bazą do zrekodowania.}

\item{key_list}{lista zawierająca .}
}
\value{
Data.frame, w którym dla kolumn wyspecyfikowanych w \code{key_list}
odpowiednie wartości w 'values' zostały podmienione na wartości z 'recodes'.
Kolumny nie zdefiniowane w \code{key_list} są zwracane bez zmian.
}
\description{
Rekoduje zbiór danych zgodnie z kluczem.
}

