% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/majkTools.R
\docType{data}
\name{scores_multi}
\alias{scores_multi}
\title{Wyniki 35 uczniów otrzymane z ConQuesta 2 (wielowymiarowe).}
\format{Character zawierający 35*8 linii z wynikami.}
\usage{
scores_multi
}
\description{
Wyniki wygenerowane przez ConQuest 2 za pomocą polecenia \code{show cases} i
wczytane za pomocą funkcji \code{\link[base]{readLines}}. Wyniki zawierają
pięć PV, oszacowanie EAP i odchylenie standardowe dla rozkładu a posteriori
dla 35 uczniów. Przy wyliczaniu wyników założono wielowymiarowy model
pomiarowy.
}
\keyword{datasets}

