% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/showNotReached.R
\name{showNotReached}
\alias{showNotReached}
\title{Zamienia brak odpowiedzi na not-reached.}
\usage{
showNotReached(dataframe, tasks = NULL, missing_code = 9,
  replace_with = "N")
}
\arguments{
\item{dataframe}{data.frame z zakodowanym testem. W kolumnach z zadaniami nie
powinno być braków danych.}

\item{tasks}{character z określeniem nazw kolumn z zadaniami. Może
stanowić także wyrażenie regularne. Jeżeli nieokreślone (domyślnie),
wszystkie kolumny traktowane są jako zadania.}

\item{missing_code}{kod brakującej odpowiedzi. Domyślnie 9 (integer).}

\item{replace_with}{kod, który ma być wprowadzony, gdy odpowiedź zostanie
 określona jako not-reached. Domyślnie "N".}
}
\description{
Zamienia brak odpowiedzi na not-reached.
}

