% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeKey.R
\name{writeKey}
\alias{writeKey}
\title{Zapisuje listę z kluczem do pliku.}
\usage{
writeKey(key_list, file)
}
\arguments{
\item{key_list}{lista z kluczem.}

\item{file}{character z nazwą pliku docelowego.}
}
\description{
Zapisuje listę z kluczem do pliku.
}

