% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recodeData.R
\name{recodeData}
\alias{recodeData}
\title{Rekoduje zbiór danych zgodnie z kluczem.}
\usage{
recodeData(raw_df, key_list)
}
\arguments{
\item{raw_df}{data.frame z bazą do zrekodowania.}

\item{key_list}{lista z kluczem.}
}
\value{
Data.frame, w którym dla kolumn wymienionymi w
\code{key_list} odpowiednie wartości w kolumnie 'value' zostały podmienione
na wartości z kolumny 'recode'. Kolumny nie wymienione w \code{key_list} są
zwracane bez zmian.
}
\description{
Rekoduje zbiór danych zgodnie z kluczem. Klucz powinien być nazwaną listą.
Nazwy elementów \code{key_list} powinny odpowiadać nazwom kolumn w
\code{raw_df}. Dla każdej kolumny, która ma być zrekodowana, definicja
wartości wyjściowych oraz docelowych powinna przyjmować postać w data.frame,
z kolumnami o nazwach 'value' (wartości wyjściowe) i 'recode'
(wartości docelowe).
}
\details{
Podczas wyszukiwania i nadpisywania wartości z klucza usuwa spacje z początku
i końca w przeszukiwanej kolumnie oraz z wartości wyjściowych i wartości
docelowych.
}

