% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumPart.R
\name{sumPart}
\alias{sumPart}
\title{Dzieli pliki z summary dla modeli z TAM}
\usage{
sumPart(sum_file, chunk = "all", sep = "---")
}
\arguments{
\item{sum_file}{character określający lokalizację pliku z \code{summary},
character z wczytanym plikiem (np. za pomocą \code{\link[base]{readLines}}) lub
bezpośrednie wywołanie \code{\link[TAM:summary.tam.mml]{summary}} na obiekcie klasy
\code{tam.mml}.}

\item{chunk}{integer z określeniem, które części zwrócić lub "all"
(domyślnie), żeby zwrócić wszystkie.}

\item{sep}{character z określeniem separatora. Jeżeli określono kilka
 separatorów, dzieli plik po nich wszystkich. Domyślnie "---" (używany w
 podsumowaniu modeli w pakiecie \code{TAM}).}
}
\description{
Dzieli pliki z summary dla modeli z TAM na części i wyrzuca do
konsoli wybrane części.
}
\examples{
\dontrun{
# Wywołanie sumPart na pliku, wszystkie części
sumPart("plik_z_summary.txt")
# Wywołanie sumPart na obiekcie z wczytanym wcześniej plikiem, 2. część
plik_sum = readLines("plik_z_summary.txt")
sumPart(plik_sum, 2)
# Wywołanie sumPart bezpośrednio na summary(), od 1. do 4. części
sumPart(summary(model_z_tam), 1:4)}
}

