% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumPart.R
\name{sumPart}
\alias{sumPart}
\title{Dzieli pliki z summary dla modeli z TAM}
\usage{
sumPart(sum_file, chunk = "all", sep = "---")
}
\arguments{
\item{sum_file}{character określający lokalizację pliku z \code{summary}.}

\item{chunk}{integer z określeniem, które części zwrócić lub "all"
(domyślnie), żeby zwrócić wszystkie.}

\item{sep}{character z określeniem separatora. Jeżeli określono kilka
 separatorów, dzieli plik po nich wszystkich. Domyślnie "---".}
}
\description{
Dzieli pliki z summary dla modeli z TAM na części i wyrzuca do
konsoli wybrane części.
}

